% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractParestECArunAgedist.R
\name{ExtractParestECArunAgedist}
\alias{ExtractParestECArunAgedist}
\title{ExtractParestECArunAgedist}
\usage{
ExtractParestECArunAgedist(fit)
}
\arguments{
\item{fit}{Output (list object) from a fitted ECA model with elements: "age"       "lga"       "hsz"       "nonlin"    "lga.cc"    "nonlin.cc" "wgl"       "wgl.cc"    "split"     "k"}
}
\value{
A list (parameters for computing the age distribution)
}
\description{
Extract parameters from a fitted ECA model (fit object); to be used for computing the age distribution
}
\author{
Ingunn Fride Tvete
}
