% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractParestECArunLGA.R
\name{ExtractParestECArunLGA}
\alias{ExtractParestECArunLGA}
\title{ExtractParestECArunLGA}
\usage{
ExtractParestECArunLGA(fit)
}
\arguments{
\item{fit}{Output (list object) from a fitted ECA model with elements: "age"       "lga"       "hsz"       "nonlin"    "lga.cc"    "nonlin.cc" "wgl"       "wgl.cc"    "split"     "k"}
}
\value{
A list (parameters for computing length given age)
}
\description{
ExtractParestECArunLGA: parameters for length given age
}
\author{
Ingunn Fride Tvete
}
