% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractParestECArunWGL.R
\name{ExtractParestECArunWGL}
\alias{ExtractParestECArunWGL}
\title{ExtractParestECArunWGL}
\usage{
ExtractParestECArunWGL(fit)
}
\arguments{
\item{fit}{Output (list object) from a fitted ECA model with elements: "age"       "lga"       "hsz"       "nonlin"    "lga.cc"    "nonlin.cc" "wgl"       "wgl.cc"    "split"     "k"}
}
\value{
A list (parameters for computing weigth given length)
}
\description{
ExtractParestECArunWGL: parameters for weigth given length
}
\author{
Ingunn Fride Tvete
}
