% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindExpectedWeight.R
\name{FindExpectedWeight}
\alias{FindExpectedWeight}
\title{FindExpectedWeight}
\usage{
FindExpectedWeight(data, fit, Par_ECA, pA)
}
\arguments{
\item{data}{Sale notes data object, with one line per landing and columns: "Boat"   "S"      "G"      "R"      "Weight" "L"      "Boat2"  "LwiB" (Boat = orig. boat numer, S = season, G = gear, R = region (ECA-region 1:9), Weight = total weight (converted to gram), L = landing, Boat2 = boat number within {S,G,R}, LwiB = landing number within boat}

\item{fit}{Output (list object) from a fitted ECA model with elements: "age"       "lga"       "hsz"       "nonlin"    "lga.cc"    "nonlin.cc" "wgl"       "wgl.cc"    "split"     "k"}

\item{Par_ECA}{Sampled parameters,  the fitted ECA model (ceta_boat_matrix, ceta_unit_matrix, ceta_cell_matrix)}

\item{pA}{Age distibution matrix}
}
\value{
A vector of expected weights (one for each landing, in gram)
}
\description{
Expected weigth in a given landing in sale notes data
}
\author{
Ingunn Fride Tvete
}
