% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindpAmatrix.R
\name{FindpAmatrix}
\alias{FindpAmatrix}
\title{FindpAmatrix}
\usage{
FindpAmatrix(data, fit, Par_ECA)
}
\arguments{
\item{fit}{Output (list object) from a fitted ECA model with elements: "age"       "lga"       "hsz"       "nonlin"    "lga.cc"    "nonlin.cc" "wgl"       "wgl.cc"    "split"     "k"}

\item{Par_ECA}{Sampled parameters,  the fitted ECA model (ceta_boat_matrix, ceta_unit_matrix, ceta_cell_matrix)}

\item{dataSale}{notes data object, with one line per landing and columns: "Boat"   "S"      "G"      "R"      "Weight" "L"      "Boat2"  "LwiB" (Boat = orig. boat numer, S = season, G = gear, R = region (ECA-region 1:9), Weight = total weight (converted to gram), L = landing, Boat2 = boat number within {S,G,R}, LwiB = landing number within boat}
}
\value{
A matrix (number of rows are equal to the number of landings in the sale notes data), each row is the age distribution for the given landing
}
\description{
FindpAmatrix: find age probability distribution matrix
}
\author{
Ingunn Fride Tvete
}
