% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GearConverteSalesnoteToECAcode.R
\name{GearConverteSalesnoteToECAcode}
\alias{GearConverteSalesnoteToECAcode}
\title{RegionConverteSalesnoteToECAcode}
\usage{
GearConverteSalesnoteToECAcode()
}
\value{
A data frame with two columns; column 1: sale notes gear, column 2: corresponding ECA-gear 1:4 (the sale notes gear is based on gear used in 2015/16)
}
\description{
GearConverteSalesnoteToECAcode: Gear; sale notes data and ECA code
}
\author{
Ingunn Fride Tvete
}
