% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LengthGivenAge.R
\name{LengthGivenAge}
\alias{LengthGivenAge}
\title{LengthGivenAge}
\usage{
LengthGivenAge(data, beta_const, beta_season, beta_gear, beta_1, epsilon_region,
  epsilon_boat, epsilon_cell, epsilon_landing, tau_epsilon_fish)
}
\arguments{
\item{data}{Sale notes data object, with one line per landing and columns: "Boat"   "S"      "G"      "R"      "Weight" "L"      "Boat2"  "LwiB" (Boat = orig. boat numer, S = season, G = gear, R = region (ECA-region 1:9), Weight = total weight (converted to gram), L = landing, Boat2 = boat number within {S,G,R}, LwiB = landing number within boat}

\item{beta_const}{ECA-parameter}

\item{beta_season}{ECA-parameter}

\item{beta_gear}{ECA-parameter}

\item{beta_1}{ECA-parameter}

\item{epsilon_region}{ECA-parameter}

\item{epsilon_boat}{ECA-parameter}

\item{epsilon_cell}{ECA-parameter}

\item{epsilon_landing}{ECA-parameter}

\item{tau_epsilon_fish}{ECA-parameter}
}
\value{
A vector of fish lengths (cm)
}
\description{
Sample the length of fish for a given age
}
\author{
Ingunn Fride Tvete
}
