% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionConverteSalesnoteToECAcode.R
\name{RegionConverteSalesnoteToECAcode}
\alias{RegionConverteSalesnoteToECAcode}
\title{RegionConverteSalesnoteToECAcode}
\usage{
RegionConverteSalesnoteToECAcode()
}
\value{
A data frame with two columns; column 1: sale notes region 0:81, column 2: corresponding ECA-region 1:9 (the sale notes regions are based on the regions from 2015/16)
}
\description{
RegionConverteSalesnoteToECAcode: Region: sale notes data and ECA code
}
\author{
Ingunn Fride Tvete
}
