% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleParFromFittedDistECA.R
\name{SampleParFromFittedDistECA}
\alias{SampleParFromFittedDistECA}
\title{SampleParFromFittedDistECA}
\usage{
SampleParFromFittedDistECA(data, fit, n_neighbours, cmat)
}
\arguments{
\item{data}{Sale notes data object, with one line per landing and columns: "Boat"   "S"      "G"      "R"      "Weight" "L"      "Boat2"  "LwiB" (Boat = orig. boat numer, S = season, G = gear, R = region (ECA-region 1:9), Weight = total weight (converted to gram), L = landing, Boat2 = boat number within {S,G,R}, LwiB = landing number within boat}

\item{fit}{Output (list object) from a fitted ECA model with elements: "age"       "lga"       "hsz"       "nonlin"    "lga.cc"    "nonlin.cc" "wgl"       "wgl.cc"    "split"     "k"}

\item{n_neighbours}{Number of neighbours}

\item{cmat}{Spatial matrix}
}
\value{
A list of sampled parameters (form the fitted ECA model)
}
\description{
Sample random parameters from fitted ECA-model
}
\author{
Ingunn Fride Tvete
}
