% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateFishery.R
\name{SimulateFishery}
\alias{SimulateFishery}
\title{SimulateFishery}
\usage{
SimulateFishery(salenotesdata, A, ages, fit)
}
\arguments{
\item{salenotesdata}{Sale notes data object, with one line per landing and columns: "Boat"   "S"      "G"      "R"      "Weight" "L"      "Boat2"  "LwiB" (Boat = orig. boat numer, S = season, G = gear, R = region (ECA-region 1:9), Weight = total weight (converted to gram), L = landing, Boat2 = boat number within {S,G,R}, LwiB = landing number within boat}

\item{A}{The Number of age groups}

\item{ages}{The Ages (range, e.g. 1:19)}

\item{fit}{Output (list object) from a fitted ECA model with elements: "age"  "lga"       "hsz"       "nonlin"    "lga.cc"    "nonlin.cc" "wgl"       "wgl.cc"    "split"     "k"}
}
\value{
A list (fishery data object, with elements "Boat"  "S"   "G" "R"  "Weight" "L"  "Boat2" "LwiB"  "ExpectedWeight" "Nofish" "ages"  "LGA"   "WGL"), LGA = length given age and WGL = weight given length
}
\description{
Simulating fishery
}
\author{
Ingunn Fride Tvete
}
