% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightGivenLength.R
\name{WeightGivenLength}
\alias{WeightGivenLength}
\title{WeightGivenLength}
\usage{
WeightGivenLength(data, length_g_age, delta_const, delta_season, delta_gear,
  delta_1, nu_region, nu_boat, nu_cell, nu_landing, tau_nu_fish)
}
\arguments{
\item{data}{Sale notes data object, with one line per landing and columns: "Boat"   "S"      "G"      "R"      "Weight" "L"      "Boat2"  "LwiB" (Boat = orig. boat numer, S = season, G 0 gear, R = region (ECA-region 1:9), Weight = total weight (converted to gram), L = landing, Boat2 = boat number within {S,G,R}, LwiB = landing number within boat}

\item{length_g_age}{Length of each fish (given the age)}

\item{delta_const}{ECA parameter}

\item{delta_season}{ECA parameter}

\item{delta_gear}{ECA parameter}

\item{delta_1}{ECA parameter}

\item{nu_region}{ECA parameter}

\item{nu_boat}{ECA parameter}

\item{nu_cell}{ECA parameter}

\item{nu_landing}{ECA parameter}

\item{tau_nu_fish}{ECA parameter}
}
\value{
A vector of fish weights (in gram)
}
\description{
Sample the weight of fish for a givenlength
}
\author{
Ingunn Fride Tvete
}
